HDH-244-add-self-paced-learning-as-type-on-hdh
