+++ src/components/University/CertCard.module.scss
@@ -47,7 +47,7 @@
     }
     p {
       position: absolute;
-      left: 8px;
+      left: 5px;
       top: 40px;
       font-size: 16px;
       transform: rotate(45deg);
