+++ src/components/University/data/sptData.ts
@@ -1,0 +1,16 @@
+import { type, IltCardItem, tileType, cardType } from "../Card";
+import { MODULES } from "../../../constants";
+
+export const ilt: IltCardItem[] = [
+  // {
+  //   title: "Introduction to the Harness Platform",
+  //   module: MODULES.platform,
+  //   type: type.user,
+  //   description:
+  //     "Self-paced, prerequisite course to all module-specific ILT courses.",
+  //   version: "Enterprise Plan of any module",
+  //   link: "https://university-registration.harness.io/introduction-to-the-harness-platform",
+  //   tileType: tileType.preReq,
+  //   cardType: cardType.SPT,
+  // },
+];
